requirements.txt
pandas
numpy
matplotlib
pyyaml
textract
